Testing V0.1
Testing of the new coil windings is underway. So far we are seeing zero effect.

The coil seems to be acting as a short circuit.

We believe this could be due to insufficient isolation of the B field due to the changes we made to the C core and magnetic dipole across the juncture. (Airgaps etc) OR insufficient energetic magnitude of the carrier wave which can be corrected with a power amplification preceding the E field transformer.

These are the next potential areas to look into our development.

In terms of test code / waveform generators. - We have begun updating our test code with new waveform files as well as adopting some of arduino's open source waveform generator into our test code to try different waveforms.

*NOTE* from Arduino's documentation in their waveformgenerator.ino - I believe the max freq we can achieve through the Arduino board is around 170Hz. We can check and verify this through some rough calculations using arduino's core clock freq later if need be. But it gives us a relative idea of the freq magnitudes we have to work with in testing at this point in time with this platform.

Tomorrow we will begin finishing up the code file of waveform generation - while also looking into changing the carrier wave magnitudes.