V0.1 Testing, AD2, Variable Waveforms
The Arduino platform is currently being abandoned due to lack of an digital to analog converter (DAC) on our R3
WHICH IS GREAT

Because now we will have a much quicker development time when it comes to the carrier wave as we are using the Analog Discovery 2 (AD2) now to not only act as our oscilloscope but also to act as our carrier wave generator since the waveforms software allows for much more ease and freedom in creating test waveforms.

Variable waveforms were tested and observed today across the E field transformer; though no over unity effect is being observed.

Tomorrow we will begin with experimenting on the orientations of the core and magnetic dipole themselves & will most likely end with creating an epoxy resin casting in order to isolate problematic airgaps in regards to the magnetic field del swirling isolation problem as mentioned by Lt. Bearden (deceased).