Version 1.0 under test
Arduino is being used as a variable pulse carrier wave signal while also providing the logic for switching our "FLYBACK" wave transistor.

Issues are being resolved in regards to the wave following path of least resistance.

Solution for feedback additives signal is being investigated at this time.

(IN PROGRESS)

If self sustained switching seems to not have any immediate solution; we can go back to gain staging the V0 MEG derivative or we can move on to V2 in trying different kind of A field reverberators.