V2.0 series testing begins
V2.0 was made and the 2. series testing began so far with effects that still obey conservation of energy.

No such effects can be seen (as of today) that might be utilized for over unity energy generation.

Alternate orientations of the core, coils, dipole, & carrier wave(s) will continue to be tested for the V2 series.

If negligible effects are seen - voltage gain staging might be the needed remedy in order to increase the magnetic curl coming off of one coil so it may be picked up by the other thru magnetic inductance.

^ This seems to be the most obvious problem area as of this moment.