V2 Testing Physics & Development Direction
Today we talked about SCALAR energy and the hypothesis for a collision of B field curl being needed in order to result in a SCALAR excitation of energy to perterb the A or phi Magnetic or Electric infinite energetic field potentials.

If this hypothesis is supported we should see an excess of energy being picked up on by the output of our coils due to the infinite energetic field potentials compressing back our SCALAR excitation which will then induce magnetic energy back into the coils as electric field energy.

Our goal now is to begin amplifying the voltage across our carrier waveform to achieve such a SCALAR perterberance.

This is hypothesized based off of the assumption that the A and Phi field potentials like any  other quantum field potential is SCALAR in nature.

For more information on the physics / diagrams needed to understand this idea / hypothesis please check out the Occulted series on https://www.youtube.com/@RedEyesBadDragon OR simply watch Dev Stream {010} "today's recording".