V4.0 & SELF POWERED FOR 1 SECOND
We have doubled our C core and turned it into an I core for Version 4.0

So far in testing we have seen self powering for the duration of our capacitor discharge 1 second!!! which is an improvement over our V3 models.

Next we will begin altering the winding ratios of the I core in order to try and achieve a better energetic ratio

(i.e. we are not sending enough energy out of the I core to both charge the cap while self powering - we are however sending enough energy into the cap in order to turn off our AD2 Vcc supply voltage to then see it being powered by the Capacitor tank for the duration of its discharge. . . again the carrier wave is not sending enough energy to cancel out the energy needed in order to send it in the first place)

This is the crux of over unity and we are approaching the solution

. . .
different core / coil orientations will also be tried in order to remedy this energetic ratio of loss to one of gain and over unity

Namaste
-V